OGM-1012 Rename existing Neo4j integration test class names and package
