+++ integrationtest/src/test/java/org/hibernate/ogm/test/integration/neo4j/embedded/EmbeddedNeo4jJtaModuleMemberRegistrationIT.java
@@ -4,10 +4,13 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.ogm.test.integration.neo4j;
+package org.hibernate.ogm.test.integration.neo4j.embedded;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
+
+import java.io.File;
+import java.io.InputStream;
 
 import org.hibernate.ogm.datastore.neo4j.Neo4j;
 import org.hibernate.ogm.datastore.neo4j.Neo4jProperties;
@@ -33,12 +36,12 @@ import org.junit.runner.RunWith;
  * @author Guillaume Scheibel &lt;guillaume.scheibel@gmail.com&gt;
  */
 @RunWith(Arquillian.class)
-public class Neo4jResourceLocalModuleMemberRegistrationIT extends ModuleMemberRegistrationScenario {
+public class EmbeddedNeo4jJtaModuleMemberRegistrationIT extends ModuleMemberRegistrationScenario {
 
 	@Deployment
 	public static Archive<?> createTestArchive() throws Exception {
 		return new ModuleMemberRegistrationDeployment
-				.Builder( Neo4jResourceLocalModuleMemberRegistrationIT.class )
+				.Builder( EmbeddedNeo4jJtaModuleMemberRegistrationIT.class )
 				.persistenceXml( persistenceXml() )
 				.manifestDependencies( "org.hibernate.ogm:${hibernate-ogm.module.slot} services, org.hibernate.ogm.neo4j:${hibernate-ogm.module.slot} services" )
 				.createDeployment();
@@ -49,18 +52,27 @@ public class Neo4jResourceLocalModuleMemberRegistrationIT extends ModuleMemberRe
 				.version( "2.0" )
 				.createPersistenceUnit()
 				.name( "primary" )
-				.transactionType( PersistenceUnitTransactionType._RESOURCE_LOCAL )
+				.transactionType( PersistenceUnitTransactionType._JTA )
 				.provider( HibernateOgmPersistence.class.getName() )
 				.getOrCreateProperties();
 		PersistenceDescriptor persistenceDescriptor = propertiesContext
 				.createProperty().name( Neo4jProperties.DATASTORE_PROVIDER ).value( Neo4j.DATASTORE_PROVIDER_NAME ).up()
 				.createProperty()
 				.name( Neo4jProperties.DATABASE_PATH )
-				.value( Neo4jJtaModuleMemberRegistrationIT.neo4jFolder() )
+				.value( neo4jFolder() )
 				.up()
 				.createProperty().name( "hibernate.search.default.directory_provider" ).value( "ram" ).up()
 				.up().up();
 		return persistenceDescriptor;
+	}
+
+	static String neo4jFolder() throws Exception {
+		java.util.Properties properties = new java.util.Properties();
+		try ( InputStream propertiesStrem = Thread.currentThread().getContextClassLoader().getResourceAsStream( "neo4j-test.properties" ) ) {
+			properties.load( propertiesStrem );
+		}
+		String buildDirectory = properties.getProperty( "build.directory" );
+		return buildDirectory + File.separator + "NEO4J_DB" + File.separator + System.currentTimeMillis();
 	}
 
 	@Override
