+++ integrationtest/src/test/java/org/hibernate/ogm/test/integration/neo4j/embedded/EmbeddedNeo4jResourceLocalModuleMemberRegistrationIT.java
@@ -4,13 +4,10 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.ogm.test.integration.neo4j;
+package org.hibernate.ogm.test.integration.neo4j.embedded;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
-
-import java.io.File;
-import java.io.InputStream;
 
 import org.hibernate.ogm.datastore.neo4j.Neo4j;
 import org.hibernate.ogm.datastore.neo4j.Neo4jProperties;
@@ -36,12 +33,12 @@ import org.junit.runner.RunWith;
  * @author Guillaume Scheibel &lt;guillaume.scheibel@gmail.com&gt;
  */
 @RunWith(Arquillian.class)
-public class Neo4jJtaModuleMemberRegistrationIT extends ModuleMemberRegistrationScenario {
+public class EmbeddedNeo4jResourceLocalModuleMemberRegistrationIT extends ModuleMemberRegistrationScenario {
 
 	@Deployment
 	public static Archive<?> createTestArchive() throws Exception {
 		return new ModuleMemberRegistrationDeployment
-				.Builder( Neo4jJtaModuleMemberRegistrationIT.class )
+				.Builder( EmbeddedNeo4jResourceLocalModuleMemberRegistrationIT.class )
 				.persistenceXml( persistenceXml() )
 				.manifestDependencies( "org.hibernate.ogm:${hibernate-ogm.module.slot} services, org.hibernate.ogm.neo4j:${hibernate-ogm.module.slot} services" )
 				.createDeployment();
@@ -52,27 +49,18 @@ public class Neo4jJtaModuleMemberRegistrationIT extends ModuleMemberRegistration
 				.version( "2.0" )
 				.createPersistenceUnit()
 				.name( "primary" )
-				.transactionType( PersistenceUnitTransactionType._JTA )
+				.transactionType( PersistenceUnitTransactionType._RESOURCE_LOCAL )
 				.provider( HibernateOgmPersistence.class.getName() )
 				.getOrCreateProperties();
 		PersistenceDescriptor persistenceDescriptor = propertiesContext
 				.createProperty().name( Neo4jProperties.DATASTORE_PROVIDER ).value( Neo4j.DATASTORE_PROVIDER_NAME ).up()
 				.createProperty()
 				.name( Neo4jProperties.DATABASE_PATH )
-				.value( neo4jFolder() )
+				.value( EmbeddedNeo4jJtaModuleMemberRegistrationIT.neo4jFolder() )
 				.up()
 				.createProperty().name( "hibernate.search.default.directory_provider" ).value( "ram" ).up()
 				.up().up();
 		return persistenceDescriptor;
-	}
-
-	static String neo4jFolder() throws Exception {
-		java.util.Properties properties = new java.util.Properties();
-		try ( InputStream propertiesStrem = Thread.currentThread().getContextClassLoader().getResourceAsStream( "neo4j-test.properties" ) ) {
-			properties.load( propertiesStrem );
-		}
-		String buildDirectory = properties.getProperty( "build.directory" );
-		return buildDirectory + File.separator + "NEO4J_DB" + File.separator + System.currentTimeMillis();
 	}
 
 	@Override
